WEB-4460 organize filters into specific directory
