+++ __tests__/unit/app/pages/clinicworkspace/components/filters/SiteFilterDropdown.test.js
@@ -6,7 +6,7 @@ import configureStore from 'redux-mock-store';
 import { thunk } from 'redux-thunk';
 import { MemoryRouter } from 'react-router-dom';
 
-import SiteFilterDropdown from '@app/pages/clinicworkspace/components/SiteFilterDropdown';
+import SiteFilterDropdown from '@app/pages/clinicworkspace/components/filters/SiteFilterDropdown';
 import { trackMetric as mockTrackMetric } from '../../../../../app/core/metricUtils';
 import useIsClinicAdmin from '@app/pages/clinicworkspace/useIsClinicAdmin';
 import { SPECIAL_FILTER_STATES } from '@app/pages/clinicworkspace/useClinicPatientsFilters';
