+++ app/pages/clinicworkspace/TideDashboardV2/filters/FilterBySites.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { setOffset } from '../tideDashboardSlice';
 import { setClinicSitesFilter } from '../tideDashboardFiltersSlice';
-import SiteFilterDropdown from '../../components/SiteFilterDropdown';
+import SiteFilterDropdown from '../../components/filters/SiteFilterDropdown';
 
 const FilterBySites = () => {
   const dispatch = useDispatch();
