+++ app/pages/clinicworkspace/TideDashboardV2/filters/FilterByTags.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { setOffset } from '../tideDashboardSlice';
 import { setPatientTagsFilter } from '../tideDashboardFiltersSlice';
-import TagFilterDropdown from '../../components/TagFilterDropdown';
+import TagFilterDropdown from '../../components/filters/TagFilterDropdown';
 
 const FilterByTags = () => {
   const dispatch = useDispatch();
