+++ app/pages/clinicworkspace/clinicPatientsFilters/FilterByDataRecency.js
@@ -6,7 +6,7 @@ import reject from 'lodash/reject';
 
 import { lastDataFilterOptions } from '../../../core/clinicUtils';
 
-import DataRecencyFilterDropdown from '../components/DataRecencyFilterDropdown';
+import DataRecencyFilterDropdown from '../components/filters/DataRecencyFilterDropdown';
 
 const FilterByDataRecency = ({
   activeFilters = {},
